feat: Use hookable for login and logout functions
